FBTweak: Refactor to 3 protocols.
`FBTweak` is a mutable, persistent tweak. This commit refactors
`FBTweak` to be a protocol that defines a read-only tweak.
`FBEditableTweak` protocol is a tweak intended to be changed be users,
and the `FBActionTweak` protocol is a tweak that contains an action.

FBMutableTweak is a mutable tweak that conforms to `FBEditableTweak`
and FBPersistentTweak behaves like the old FBTweak and persists its
value to NSUserDefaults.

This refactor make it easy to create a classes the expose immutable
tweaks and for the UI to decide which cell to used.